+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -23,6 +23,7 @@ CHANGELOG
  * using the `PhpDumper` with an uncompiled `ContainerBuilder` is deprecated and
    will not be supported anymore in 4.0
  * deprecated the `DefinitionDecorator` class in favor of `ChildDefinition`
+ * allow config files to be loaded using a glob pattern
 
 3.2.0
 -----